docs: fix `null` type descriptions in Introspection docs
We're currently specifying descriptions in the `StarWarsSchema` using comments, which is causing them to resolve as `null` when we run introspection queries, like in the [last example on the Introspection docs page](https://graphql.org/learn/introspection/).

The [newer versions of the spec](https://spec.graphql.org/October2021/#sec-Descriptions) outline that descriptions should use a `BlockString`, so this change updates all the descriptions to use `BlockString` instead of comments.

Looks like, alternatively, we could have specified `commentDescriptions` in our invocation of `makeExecutableSchema`, but we might as well make our schema consistent with the current way of specifying descriptions in the spec.

Fixes #1496